Replace render_to_response with render in views.py
This fixes a bug in django 1.10 that meant the csrf token was not being
inserted into the edit form.

Previously the page, edit, revisions, changes, revision_list, and
page_list views used render_to_response, passing RequestContext(request)
as the third positional argument (in django 1.9 and earlier this was the
context_instance parameter).

As of django 1.10 the render_to_response shortcut function does not have
the context_instance parameter. See:

https://docs.djangoproject.com/en/1.9/_modules/django/shortcuts/#render_to_response
and
https://docs.djangoproject.com/en/1.10/_modules/django/shortcuts/#render_to_response

This meant that the request was not being passed to the template, and so
the csrf_token was not being included in any forms. Attempting to create
any wiki pages therefore resulted in a 403 csrf invalid error.

This change replaces the use of render_to_response with render which
(unlike render_to_response) passes the request object through to
render_to_string. This ensures the csrf token is added to the page.