Create a new template tag to handle creation of html
The default behaviour of this tag is to convert line breaks to <p> tags,
and to turn strings that are obviously urls into links to those urls.
This default behaviour is wrapped into the new preprocess_content
filter.

The default behaviour can be overridden by setting the
WAKAWAKA_PREPROCESS_CONTENT_FUNCTION variable in settings.py. Any
callable can be used. The only requirement is that it must take a string
and return a string.

Added tests for default and overridden preprocessor functions.

Also updated the README to mention the new setting for a custom pre
processor.